requirements.txt
absl-py==1.2.0
appnope==0.1.3
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
cvzone==1.3.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
djitellopy2==2.3
entrypoints==0.4
fonttools==4.37.4
idna==3.3
imutils==0.5.4
ipykernel==6.15.3
ipython==7.34.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
kiwisolver==1.4.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mediapipe==0.8.5
nest-asyncio==1.5.5
numpy==1.20.3
opencv-contrib-python==4.5.5.62
opencv-python==4.6.0.66
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.31
protobuf==3.17.3
psutil==5.9.2
ptyprocess==0.7.0
PyAudio==0.2.11
pycodestyle==2.9.1
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
pyzmq==24.0.1
requests==2.28.1
seaborn==0.12.0
six==1.16.0
srt==3.5.2
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
typing_extensions==4.4.0
urllib3==1.26.12
vosk==0.3.43
wcwidth==0.2.5
